test: OAuthハンドラの結合テストを実装ファイルごとに分割
github.test. tsに集約していたOAuth全ハンドラのテストを、実装ファイルと対になる
github-login/callback/link/status/unlink.test.tsへ分割（server/auth/handlerの慣習に統一）。
set-cookie検索は共通ヘルパfindSetCookieに切り出し。
