--- application/apps/web/src/server/oauth/handler/github.test.ts
@@ -1,172 +1,0 @@
-import { apiRequest } from '@/test/helper/request'
-import type { CleanUpIds } from '@/test/helper/user'
-import * as userHelper from '@/test/helper/user'
-
-function get(path: string, cookies?: string) {
-  return apiRequest(path, { method: 'GET', cookies })
-}
-
-function getSetCookies(res: Response): string[] {
-  return res.headers.getSetCookie()
-}
-
-// GitHub本体の認可画面〜コード発行は外部サービスのため通せない。ここでは認可URLの発行・
-// リダイレクト先の妥当性・認証ガードなど、アプリ側で完結する境界のみを検証する
-describe('GitHub OAuth認証', () => {
-  describe('正常系', () => {
-    it('ログイン開始でSupabaseのGitHub認可URLへリダイレクトする', async () => {
-      const res = await get('/api/oauth/github/login')
-
-      expect(res.status).toBe(302)
-      const location = res.headers.get('Location') ?? ''
-      expect(location).toContain('/auth/v1/authorize')
-      expect(location).toContain('provider=github')
-    })
-
-    it('ログイン開始でフロー種別Cookieにloginを保存する', async () => {
-      const res = await get('/api/oauth/github/login')
-
-      expect(res.status).toBe(302)
-      const flowCookie = getSetCookies(res).find((cookie) => cookie.startsWith('oauth_flow='))
-      expect(flowCookie).toContain('oauth_flow=login')
-    })
-
-    it('redirectクエリが内部パスなら戻り先Cookieに保存する', async () => {
-      const res = await get('/api/oauth/github/login?redirect=%2Fsettings')
-
-      expect(res.status).toBe(302)
-      const redirectCookie = getSetCookies(res).find((cookie) =>
-        cookie.startsWith('oauth_redirect_to='),
-      )
-      expect(redirectCookie).toContain('oauth_redirect_to=%2Fsettings')
-      expect(redirectCookie).toContain('HttpOnly')
-    })
-
-    it('redirectクエリが無ければ残存する戻り先Cookieを削除する', async () => {
-      const res = await get('/api/oauth/github/login')
-
-      expect(res.status).toBe(302)
-      const redirectCookie = getSetCookies(res).find((cookie) =>
-        cookie.startsWith('oauth_redirect_to='),
-      )
-      expect(redirectCookie).toContain('Max-Age=0')
-    })
-
-    // ログイン済みユーザーが必要なケースだけユーザーを作成する
-    describe('ログイン済みユーザーの連携管理', () => {
-      const TEST_EMAIL = 'oauth-github-test@example.com'
-      const TEST_PASSWORD = 'Test@password123'
-      const createdIds: CleanUpIds = { userIds: [], authIds: [] }
-
-      beforeEach(async () => {
-        const { userId, authenticationId } = await userHelper.create(TEST_EMAIL, TEST_PASSWORD)
-        createdIds.userIds.push(userId)
-        createdIds.authIds.push(authenticationId)
-      })
-
-      afterEach(async () => {
-        await userHelper.cleanUp(createdIds)
-        createdIds.userIds.length = 0
-        createdIds.authIds.length = 0
-      })
-
-      it('連携開始でSupabaseのGitHub認可URLへリダイレクトし、フロー種別と戻り先Cookieを保存する', async () => {
-        const { cookies } = await userHelper.login(TEST_EMAIL, TEST_PASSWORD)
-
-        const res = await get('/api/oauth/github/link', cookies)
-
-        expect(res.status).toBe(302)
-        const location = res.headers.get('Location') ?? ''
-        expect(location).toContain('/auth/v1/authorize')
-        expect(location).toContain('provider=github')
-
-        const setCookies = getSetCookies(res)
-        expect(setCookies.find((cookie) => cookie.startsWith('oauth_flow='))).toContain(
-          'oauth_flow=link',
-        )
-        expect(setCookies.find((cookie) => cookie.startsWith('oauth_redirect_to='))).toContain(
-          'oauth_redirect_to=%2Fsettings',
-        )
-      })
-
-      it('未連携の連携状態はlinked=falseを返す', async () => {
-        const { cookies } = await userHelper.login(TEST_EMAIL, TEST_PASSWORD)
-
-        const res = await get('/api/oauth/github', cookies)
-
-        expect(res.status).toBe(200)
-        expect(await res.json()).toEqual({ linked: false })
-      })
-
-      // 未連携ユーザーは早期returnで204を返すno-opパス。連携済みを前提とする実DELETE解除は
-      // OAuthフル往復が必要なため、ドメインのユニットテストで担保する
-      it('未連携での解除は何もせず204を返す', async () => {
-        const { cookies } = await userHelper.login(TEST_EMAIL, TEST_PASSWORD)
-
-        const res = await apiRequest('/api/oauth/github', {
-          method: 'DELETE',
-          cookies,
-          contentTypeJson: true,
-        })
-
-        expect(res.status).toBe(204)
-      })
-    })
-  })
-
-  describe('準正常系', () => {
-    it.each([
-      {
-        name: 'codeなしの失敗はエラー種別を添えてログイン画面へ戻す',
-        path: '/api/oauth/github/callback?error=access_denied',
-        cookies: undefined,
-        location: '/login?oauthError=github',
-      },
-      {
-        name: '連携フローの失敗はログイン状態を保ったまま設定画面へ戻す',
-        path: '/api/oauth/github/callback?error=access_denied',
-        cookies: 'oauth_flow=link; oauth_redirect_to=%2Fsettings',
-        location: '/settings?oauthError=github',
-      },
-      {
-        name: '戻り先が設定画面でもログインフローの失敗はログイン画面へ戻す',
-        path: '/api/oauth/github/callback?error=access_denied',
-        cookies: 'oauth_flow=login; oauth_redirect_to=%2Fsettings',
-        location: '/login?oauthError=github',
-      },
-      {
-        name: '検証情報のない不正なcodeはログイン画面へ戻す',
-        path: '/api/oauth/github/callback?code=invalid-code',
-        cookies: undefined,
-        location: '/login?oauthError=github',
-      },
-    ])('$name', async ({ path, cookies, location }) => {
-      const res = await get(path, cookies)
-
-      expect(res.status).toBe(302)
-      expect(res.headers.get('Location')).toBe(location)
-    })
-
-    it('redirectクエリが外部URLなら戻り先Cookieに保存せず削除する', async () => {
-      const res = await get('/api/oauth/github/login?redirect=https%3A%2F%2Fevil.example.com%2F')
-
-      expect(res.status).toBe(302)
-      const redirectCookie = getSetCookies(res).find((cookie) =>
-        cookie.startsWith('oauth_redirect_to='),
-      )
-      expect(redirectCookie).toContain('oauth_redirect_to=;')
-      expect(redirectCookie).toContain('Max-Age=0')
-    })
-
-    it.each([
-      { name: '連携開始', method: 'GET', path: '/api/oauth/github/link' },
-      { name: '連携状態の取得', method: 'GET', path: '/api/oauth/github' },
-      { name: '連携解除', method: 'DELETE', path: '/api/oauth/github' },
-    ])('未ログインでの$nameは401を返す', async ({ method, path }) => {
-      // Content-Type未指定だとcsrf()がtext/plain扱いでDELETEを403にするため、JSONを明示する
-      const res = await apiRequest(path, { method, contentTypeJson: true })
-
-      expect(res.status).toBe(401)
-    })
-  })
-})
