+++ application/apps/web/src/test/helper/request.ts
@@ -61,3 +61,7 @@ export function apiRequest(
     env,
   )
 }
+
+export function findSetCookie(res: Response, prefix: string): string | undefined {
+  return res.headers.getSetCookie().find((cookie) => cookie.startsWith(prefix))
+}
